# FE/src/app/services/models/book-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookRequest {
  NoOfPages: string;
  author: string;
  description: string;
  id?: number;
  isbn: string;
  language: string;
  title: string;
}