# FE/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponse {
  author?: string;
  cover?: string;
  description?: string;
  id?: number;
  isbn?: string;
  language?: string;
  noOfPages?: string;
  rating?: number;
  title?: string;
}