# FE/src/app/services/models/page-response-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookResponse } from '../models/book-response';
export interface PageResponseBookResponse {
  content?: Array<BookResponse>;
  firstPage?: boolean;
  lastPage?: boolean;
  pageNumber?: number;
  sizeOfPage?: number;
  totalElements?: number;
  totalPages?: number;
}