# FE/src/app/services/models/page-response-borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedBookResponse } from '../models/borrowed-book-response';
export interface PageResponseBorrowedBookResponse {
  content?: Array<BorrowedBookResponse>;
  firstPage?: boolean;
  lastPage?: boolean;
  pageNumber?: number;
  sizeOfPage?: number;
  totalElements?: number;
  totalPages?: number;
}