# FE/src/app/services/models/page-response-review-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReviewResponse } from '../models/review-response';
export interface PageResponseReviewResponse {
  content?: Array<ReviewResponse>;
  firstPage?: boolean;
  lastPage?: boolean;
  pageNumber?: number;
  sizeOfPage?: number;
  totalElements?: number;
  totalPages?: number;
}